Add some good universal scalac options
