--- core/src/main/scala/filodb.core/Messages.scala
@@ -1,26 +1,0 @@
-package filodb.core
-
-/**
- * Common messages for inter-Actor communication
- */
-package object messages {
-  // The parent trait for all commands, esp sent down to I/O data storage actors/futures
-  trait Command
-
-  // The parent trait for all responses from the I/O data storage actors/futures
-  trait Response
-
-  // Parent trait for error messages
-  trait ErrorResponse extends Response
-
-  // Common responses
-  case class StorageEngineException(t: Throwable) extends ErrorResponse
-  case class MetadataException(t: Throwable) extends ErrorResponse
-  case object NoSuchCommand extends ErrorResponse
-  case object InconsistentState extends ErrorResponse  // for conditional updates - race condition detected!
-
-  case object NotFound extends Response
-  case object NotApplied extends Response
-  case object AlreadyExists extends Response
-  case object Success extends Response
-}
